refactor : change api endpoint urls
