+++ src/main/java/com/softgallery/story_playground_server/controller/ChatGptController.java
@@ -25,9 +25,4 @@ public class ChatGptController {
     public ChatGptResponseDTO sendQuestion(@RequestBody QuestionRequestDTO requestDTO) {
         return chatGptService.askQuestion(requestDTO);
     }
-
-//    @GetMapping("/token")
-//    public ResponseEntity<String> getToken() {
-//        return ResponseEntity.ok().body(this.gptToken);
-//    }
 }
