+++ src/main/java/com/softgallery/story_playground_server/controller/StoryListController.java
@@ -6,11 +6,12 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/main")
+@RequestMapping("/story/list")
 public class StoryListController {
     private final StoryListService storyListService;
 
@@ -18,17 +19,23 @@ public class StoryListController {
         this.storyListService = storyListService;
     }
 
-    @GetMapping("/list/incomplete")
+    @GetMapping("/incomplete")
     public ResponseEntity<List<StoryInfoDTO>> loadIncompleteStories(@RequestHeader(name = "Authorization") String userToken) {
         List<StoryInfoDTO> incompleteStories = storyListService.findIncompleteStoriesMadeByUserName(userToken);
 
         return ResponseEntity.ok().body(incompleteStories);
     }
 
-    @GetMapping("/list/complete")
+    @GetMapping("/complete")
     public ResponseEntity<List<StoryInfoDTO>> loadCompleteStories(@RequestHeader(name = "Authorization") String userToken) {
         List<StoryInfoDTO> completeStories = storyListService.findCompleteStoriesMadeByUserName(userToken);
 
         return ResponseEntity.ok().body(completeStories);
+    }
+
+    @GetMapping("/{storyId}")
+    public ResponseEntity<List<String>> loadStory(@RequestHeader(name = "Authorization") String userToken,
+                                                  @PathVariable Long storyId) {
+        return ResponseEntity.ok().body(storyListService.findStoryByStoryId(storyId));
     }
 }
