+++ src/main/java/com/softgallery/story_playground_server/controller/TestController.java
@@ -8,22 +8,24 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @ResponseBody
+@RequestMapping("/test")
 public class TestController {
-    @GetMapping("/")
+    @GetMapping("/page")
     public String home() {
         return "home";
     }
 
-    @GetMapping("/test")
+    @GetMapping("/response")
     public ResponseEntity<String> test() {
         return ResponseEntity.ok().body("auth OK");
     }
 
-    @PostMapping("/test/moderation")
+    @PostMapping("/moderation")
     public WordFilterDTO testPost(@RequestBody UserDTO userDTO) {
         return WordFilter.doFilterWithGptModeration(userDTO.getUsername());
     }
